moved api call of storage to platform folder
