+++ src/test/K6/src/api/platform/storage/events.js
@@ -1,6 +1,6 @@
 import http from "k6/http";
-import * as config from "../../config.js";
-import * as header from "../../buildrequestheaders.js"
+import * as config from "../../../config.js";
+import * as header from "../../../buildrequestheaders.js"
 
 //Api call to Platform:Storage to add an event to an instance
 export function postAddEvent(altinnStudioRuntimeCookie, partyId, instanceId, eventData) {
