+++ src/test/K6/src/tests/app/instances.js
@@ -7,8 +7,8 @@
 import { check } from "k6";
 import { addErrorCount } from "../../errorcounter.js";
 import * as appInstances from "../../api/app/instances.js"
-import * as platformInstances from "../../api/storage/instances.js"
-import { deleteSblInstance } from "../../api/storage/messageboxinstances.js"
+import * as platformInstances from "../../api/platform/storage/instances.js"
+import { deleteSblInstance } from "../../api/platform/storage/messageboxinstances.js"
 import * as setUpData from "../../setup.js";
 
 const userName = __ENV.username;
