+++ src/test/K6/src/tests/app/negativetests/appowner-writewithoutaccess.js
@@ -17,8 +17,8 @@ import { convertMaskinPortenToken } from "../../../api/platform/authentication.j
 import * as appInstances from "../../../api/app/instances.js"
 import * as appData from "../../../api/app/data.js"
 import * as appProcess from "../../../api/app/process.js"
-import * as apps from "../../../api/storage/applications.js"
-import * as storageInstances from "../../../api/storage/instances.js"
+import * as apps from "../../../api/platform/storage/applications.js"
+import * as storageInstances from "../../../api/platform/storage/instances.js"
 import * as setUpData from "../../../setup.js";
 
 const userName = __ENV.username;
