+++ src/test/K6/src/tests/app/negativetests/withoutallowedroles.js
@@ -7,7 +7,7 @@
 import { check } from "k6";
 import { addErrorCount } from "../../../errorcounter.js";
 import * as appInstances from "../../../api/app/instances.js"
-import * as platformInstances from "../../../api/storage/instances.js"
+import * as platformInstances from "../../../api/platform/storage/instances.js"
 import * as setUpData from "../../../setup.js";
 
 const userName = __ENV.username;
