+++ src/test/K6/src/tests/app/sirius.js
@@ -19,11 +19,11 @@
 import { check, sleep, group } from "k6";
 import { addErrorCount, printResponseToConsole } from "../../errorcounter.js";
 import * as appInstances from "../../api/app/instances.js"
-import * as apps from "../../api/storage/applications.js"
+import * as apps from "../../api/platform/storage/applications.js"
 import * as appData from "../../api/app/data.js"
 import * as appProcess from "../../api/app/process.js"
-import * as platformInstances from "../../api/storage/instances.js"
-import { deleteSblInstance } from "../../api/storage/messageboxinstances.js"
+import * as platformInstances from "../../api/platform/storage/instances.js"
+import { deleteSblInstance } from "../../api/platform/storage/messageboxinstances.js"
 import { convertMaskinPortenToken } from "../../api/platform/authentication.js"
 import * as setUpData from "../../setup.js";
