+++ src/test/K6/src/tests/app/uploadanddownload.js
@@ -22,10 +22,10 @@ import { addErrorCount, printResponseToConsole } from "../../errorcounter.js";
 import * as appInstances from "../../api/app/instances.js"
 import * as appData from "../../api/app/data.js"
 import * as appProcess from "../../api/app/process.js"
-import * as platformInstances from "../../api/storage/instances.js"
-import * as apps from "../../api/storage/applications.js"
-import * as storageData from "../../api/storage/data.js"
-import { deleteSblInstance } from "../../api/storage/messageboxinstances.js"
+import * as platformInstances from "../../api/platform/storage/instances.js"
+import * as apps from "../../api/platform/storage/applications.js"
+import * as storageData from "../../api/platform/storage/data.js"
+import { deleteSblInstance } from "../../api/platform/storage/messageboxinstances.js"
 import * as setUpData from "../../setup.js";
 
 const instanceFormDataXml = open("../../data/" + level2App + ".xml");
