+++ src/test/K6/src/tests/platform/appowner/completeconfirmation.js
@@ -11,7 +11,7 @@
 
 import { check } from "k6";
 import { printResponseToConsole } from "../../../errorcounter.js";
-import * as storageInstances from "../../../api/storage/instances.js"
+import * as storageInstances from "../../../api/platform/storage/instances.js"
 import { convertMaskinPortenToken } from "../../../api/platform/authentication.js"
 import * as setUpData from "../../../setup.js";
 import Papa from "https://jslib.k6.io/papaparse/5.1.1/index.js";
