+++ src/test/K6/src/tests/platform/appowner/getinstances5482.js
@@ -12,7 +12,7 @@
 import { check } from "k6";
 import { addErrorCount } from "../../../errorcounter.js";
 import { convertMaskinPortenToken } from "../../../api/platform/authentication.js"
-import * as instances from "../../../api/storage/instances.js"
+import * as instances from "../../../api/platform/storage/instances.js"
 
 const appOwner = "ttd";
 const app = "testcase-5482";
