+++ src/test/K6/src/tests/platform/appowner/instances.js
@@ -14,7 +14,7 @@
 import { check } from "k6";
 import { addErrorCount } from "../../../errorcounter.js";
 import { convertMaskinPortenToken } from "../../../api/platform/authentication.js"
-import * as instances from "../../../api/storage/instances.js"
+import * as instances from "../../../api/platform/storage/instances.js"
 import * as setUpData from "../../../setup.js";
 
 const userName = __ENV.username;
