+++ src/test/K6/src/tests/platform/receipt/receipt.js
@@ -7,7 +7,7 @@
 import { check } from "k6";
 import { addErrorCount } from "../../../errorcounter.js";
 import * as setUpData from "../../../setup.js";
-import * as instances from "../../../api/storage/instances.js"
+import * as instances from "../../../api/platform/storage/instances.js"
 import * as receipt from "../../../api/platform/receipt.js"
 
 const userName = __ENV.username;
