+++ src/test/K6/src/tests/platform/storage/data.js
@@ -5,10 +5,10 @@
 */
 
 import { check } from "k6";
-import * as apps from "../../../api/storage/applications.js"
-import * as instances from "../../../api/storage/instances.js"
-import * as instanceData from "../../../api/storage/data.js"
-import * as sbl from "../../../api/storage/messageboxinstances.js"
+import * as apps from "../../../api/platform/storage/applications.js"
+import * as instances from "../../../api/platform/storage/instances.js"
+import * as instanceData from "../../../api/platform/storage/data.js"
+import * as sbl from "../../../api/platform/storage/messageboxinstances.js"
 import * as setUpData from "../../../setup.js";
 import { addErrorCount } from "../../../errorcounter.js";
