+++ src/test/K6/src/tests/platform/storage/messageboxinstances.js
@@ -5,8 +5,8 @@
 */
 
 import { check } from "k6";
-import * as instances from "../../../api/storage/instances.js"
-import * as sbl from "../../../api/storage/messageboxinstances.js"
+import * as instances from "../../../api/platform/storage/instances.js"
+import * as sbl from "../../../api/platform/storage/messageboxinstances.js"
 import * as setUpData from "../../../setup.js";
 import * as support from "../../../support.js";
 import { addErrorCount } from "../../../errorcounter.js";
