+++ src/test/K6/src/tests/platform/storage/process.js
@@ -5,9 +5,9 @@
 */
 
 import { check } from "k6";
-import * as instances from "../../../api/storage/instances.js"
-import * as process from "../../../api/storage/process.js"
-import * as sbl from "../../../api/storage/messageboxinstances.js"
+import * as instances from "../../../api/platform/storage/instances.js"
+import * as process from "../../../api/platform/storage/process.js"
+import * as sbl from "../../../api/platform/storage/messageboxinstances.js"
 import * as setUpData from "../../../setup.js";
 import { addErrorCount } from "../../../errorcounter.js";
